backend/pyproject.toml
[project]
name = "family-budget-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-environ>=0.11",
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "factory-boy>=3.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
